+++ pwsh/dev/functions/getOrphanedResources.ps1
@@ -3,6 +3,13 @@ function getOrphanedResources {
     Write-Host 'Getting orphaned resources (ARG)'
 
     $queries = [System.Collections.ArrayList]@()
+    $intent = 'cost savings - stopped but not deallocated VM'
+    $null = $queries.Add([PSCustomObject]@{
+            queryName = 'microsoft.compute/virtualmachines/instanceview/stopped'
+            query     = "Resources | where type =~ 'microsoft.compute/virtualmachines' and properties.extended.instanceView.powerState.code =~ 'PowerState/stopped' | project type, subscriptionId, Resource=id, Intent='$intent'"
+            intent    = $intent
+        })
+
     $intent = 'clean up'
     $null = $queries.Add([PSCustomObject]@{
             queryName = 'microsoft.resources/subscriptions/resourceGroups'